drm/i915: Avoid accidental exposure of erroneous GGTT PTE
If the vma bind pipeline fails, we cancel writing the now invalid PTE
into the GTT, assuming that the range was safely cleared by the previous
unbind. However, with the GGTT we do not automatically replace the PTE
on unbind as we assume that nothing can access the GGTT without kernel
mediation. An erroneous vma presents a state of flux, where although we
propagate that error to all consumes of the vma, and we expect the use
of the vma to be revoked, be defensive and clear the dead range just in
case HW is still programmed to access the invalid PTE.

Suggested-by: Matthew Auld <matthew.auld@intel.com>
Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Reviewed-by: Matthew Auld <matthew.auld@intel.com>