INTEL_DII/NOTUPSTREAM: drm/i915/dmabuf: Internal p2p distance verification
Specific bridge information is needed for the pci_p2pdma_distance()
check to return useful information.

The bridge info has been up-streamed for the latest Xeon servers, but
it is possible that it has not been backported to a specific distro.

Augment the distance check, locally, in i915 to work around the
missing info.

v2: do a get/put_device to model distance code more closely
v3: reflect review comments

Reviewed-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Signed-off-by: Michael J. Ruhl <michael.j.ruhl@intel.com>