drm/i915/gt: Clear wedged status upon suspend
Currently we use set-wedged on suspend if the workload is not responding
in order to allow a fast suspend (albeit at the cost of discarding the
current userspace). This may leave the device wedged during suspend,
where we may require the device available in order to swapout CPU
inaccessible device memory. Clear any temporary wedged-status after
flushing userspace off the device so we can use the blitter ourselves
inside suspend.

Testcase: igt/gem_eio/in-flight-suspend
Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>