drm/i915: Unbind vma if asked to pin with conflicting flags
Currently, if a vma is unpinned but bound into the drm_mm and we try to
pin it with conflicting alignment or flags, we simply rebind it into the
current location despite the misplacement. Force the unbind (if allowed)
if there is a conflict between the current position of the vma in the
drm_mm and the caller's requirements.

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>