drm/i915: Replace idle_barrier llist with lock+list
When running perf [OA] with multiple processes starting in parallel,
list corruption was observed within the fence callbacks. Janusz tracked
this down to a race within the idle barrier reuse, in particular the
impact of two threads trying to reclaim their idle barrier
simultaneously in replace_barrier(). In essence, we failed to handle
the negative condition where we had an idle_barrier but the llist was
busy on the other thread. We would reuse the idle_barrier believing we
had ownership, but the second thread (in the process of claiming its own
idle_barrier) would replace the first idle_barrier back onto the
engine's llist. Thus the idle_barrier existed on two incompatible lists
simultaneously.

By replacing the engine->idle_barriers with a spinlocked list, we can
elminate this racy search and reclamation with a single list_del.

Closes: https://gitlab.freedesktop.org/drm/intel/-/issues/6333
Co-developed-by: Janusz Krzysztofik <janusz.krzysztofik@linux.intel.com>
Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>