drm/i915 : Enable Pipelined L3 Fabric Flush.
Userspace flows use various PIPE_CONTROL flushes to ensure
synchronization between workloads. One of these, the L3 fabric flush,
ensures that prior workloads are complete before allowing a subsequent
workload to proceed. It does this by inserting the flush on top of the
pipe as part of a post sync operation on a flush completion.

PIPE_CONTROL has a second mode, Depth Stall Enable, which causes
workloads and flushes to become pipelined until after the sync and post
sync operations are complete.

When both of these flushes are programmed simultaneously, the L3 fabric
flush can be issued prior to workload completion based on that
pipelining. In some scenarios, this can lead to a flush happening out of
order and a second workload seeing stale data.

There is an MMIO register bit (WM_CHICKEN2[5]) which can be used to
ensure the flushes specifically wait until after the workload completes
instead of being directly pipelined between workloads. By default, this
mode is not enabled, assuming the depth stall and l3 fabric simultaneous
programming is not default. When this is the case, the MMIO bit needs to
be programmed.

Additionally, since WM_CHICKEN2 is not accessible from userspace, the KMD
must program this bit in the golden context in the event the userspace
driver decides to utilize this programming mode.

Bspec: 46132, 46209

Cc:Stuart Summers <stuart.summers@intel.com>
Cc:Matt Roper <matthew.d.roper@intel.com>
Cc:Ayaz A Siddiqui <ayaz.siddiqui@intel.com>
Signed-off-by: Himal Prasad Ghimiray <himal.prasad.ghimiray@intel.com>
Reviewed-by: Stuart Summers <stuart.summers@intel.com>
Acked-by: Matt Roper <matthew.d.roper@intel.com>