INTEL_DII: drm/i915/gem: Dynamically tune lmem clear chunk size
In principle, we want to emit large contiguous clears to maximise the
blitter bandwidth while maintaining a fine enough arbitration interval
to allow for low latency preemption between competing timeslices. The
penalty for exceeding the execution quantum is context termination
(reset), which for the blitter clears will cause cascading failures.

We pick a small target duration of 1ms to be safe for all scheduler
backends and under different frequency profiles, and then measure a
blitter clear operation to estimate the maximum chunk size that can be
executed within that duration.

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>