drm/i915/debugfs: Allow for longer grace periods in quiescing the system
Use the preemption timeout and heartbeat interval as a guide for how
long we can expect the system to flush the last remaining requests
before forcing a reset.

The initial intent was to only wait a short while after IGT and force a
reset if there was still work pending (assumed to be a broken test or
driver). Using a short timeout would ensure that even a broken IGT would
still complete within a reasonable timeframe. However, not all HW is
equal and we would like a little more discretion before forcing the
reset, so look at the configuration values for the preemption timeout
and heartbeat and use those as a guide for how fast we can expect to
flush the engine's workqueues.

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Acked-by: Jonathan Cavitt <jonathan.cavitt@intel.com>