INTEL_DII: drm/i915/vf: Disable status page sanitization during fixups
We want to save any requests being executed during migration. To do
that, we need information about when the execution ended on the ring -
which is within status page of each context. But during gt sanitization
which is part of resume code, the status pages are assumed dirty and
therefore re-initialized.

This patch disables status page sanitization callback during the
post-migration worker, so that the data within pages is preserved,
and fixups can be applied to it.

Signed-off-by: Tomasz Lis <tomasz.lis@intel.com>
Acked-by: Michal Wajdeczko <michal.wajdeczko@intel.com>