INTEL_DII: drm/i915/gem: Export direct lmem clears for objects
Allow callers to perform a fast blitter clear on lmem objects. The
blitter operation is submitted to the blitter HW queue with the
completion fence stored on the object, returning to the caller before
the fence is signaled. Should the caller need to know the result of the
clear, they may wait upon i915_gem_object_migrate_sync().

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Acked-by: Nirmoy Das <nirmoy.das@intel.com>