drm/i915/gt: Fix active request tracking
Commit eb5e7da736f3 introduced a new list for tracking active requests
local to contexts, duplicating the existing list we have for tracking
requests along contexts and removing the common code for tracking active
requests along the [scheduling] engines. This has the consequence, not
only of code duplication, but that the local active context tracking is
not accurate for handling wedged request cancellation, leaving requests
dangling, lost to the driver and the system unable to recover. By
reinstating the common code again, we have both per-context request
unwinding and global active request tracking, along with correct locking
for retreiving those active requests.

Fixes: eb5e7da736f3 ("drm/i915/guc: Reset implementation for new GuC interface")
Fixes: dc0dad365c5e ("drm/i915/guc: Fix for error capture after full GPU reset with GuC")
Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>