INTEL_DII: drm/i915/pvc: Mark early steppings as pre-production
PVC with A-step base die and/or A-step compute tile are pre-production
platforms and are not officially supported.

FIXME:  We'll actually remove the A-step-specific driver code in the
future; for now we just want to taint the kernel and let this soak for a
while to make sure everyone has truly moved off this old hardware.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>