INTEL_DII: drm/i915/gem: Touch and clear all available device memory
Sometimes we need to touch or clear all of device memory in order to
verify correct access, or to scrub stale content. We can do this from
userspace with a trial and error approach to pinning all available
device memory and clearing the pages, but we can do it much more
efficiently from within the kernel where we can iterate directly over
the available buddy blocks and write straight to device memory.

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Cc: Stuart Summers <stuart.summers@intel.com>