drm/i915/gt: Sanitycheck array bounds for xehp_regs[] lookup
As we index xehp_regs[] by the engine->class, it is not clear to static
analysers that the lookup is always well defined. Add a sanity check,
similar to the older platforms, to verify that the array index is valid
and throw a warning should this change for future platforms (without a
corresponding update).

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Cc: Matt Roper <matthew.d.roper@intel.com>