INTEL_DII: drm/i915: Refactoring of i915_gem_map_dma_buf
In preparation to support segmented (chunked) BOs, i915_gem_copy_pages()
is merged back into i915_gem_map_dma_buf().  By doing so, the calls to
do the dma mapping are right after iterating over the SGL, and this code
becomes more efficient in follow-on patch which adds support for
segmented BOs, where we need to individually map each segment and
where we can have a mix of different backing store types in the SGT.

Additional changes:
- i915_gem_map_dma_buf is renamed to i915_gem_copy_map_dma_buf.
- update map_fabric_connectivity() to return an error code instead of
NULL pointer on error.
- dmabuf_[map|unmap]_addr() are updated to do the dma
mapping for either SMEM or LMEM objects instead of testing for that
in the caller.

Reviewed-by: Michael J. Ruhl <michael.j.ruhl@intel.com>
Signed-off-by: Brian Welty <brian.welty@intel.com>