drm/i915/gt: Reuse the next kernel request as the next heartbeat
In a few cases, such as during context close, we may issue more requests
along the kernel context to trigger preemption of active clients. These
requests may also function as heartbeats, indicators of forward
progress, so if we see that we still have pending requests queued adopt
the earlier as the next heartbeat to reduce the amount of work we
generate on an already heavily subscribed engine.

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>