drm/i915/gt: Ratelimit preemption pulses
Upon context close we may want to send a preemptive pulse down the
engine to remove th active context from the hardware. We only need to
send a new pulse if we don't have one queued (but not yet executing), so
first check for an existing pulse we can piggyback.

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>