INTEL_DII: drm/i915/gt: Wait for the engine to stall between heartbeats
While the engine is active processing userspace batches, there is a
continual flow of interrupts - between every request we generate an
interrupt, the user batch itself may trigger an interrupt, and upon a
context switch there will be more interrupts. Therefore, while there is
forward progress we can be reasonably sure there will be interrupts, and
only if we haven't seen an interrupt for a period, do we need to kick
the system with a heartbeat.

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>