INTEL_DII: drm/i915/selftests: Flush PTE write before read in iov_ggtt_live
Starting from MTL, we observe a delay in the writing of PTE MMIO, which
causes a problem when reading this value immediately after write in
iov_ggtt_live selftest.
Based on HSDES discussion, the solution to this problem is to add an
additional MMIO read after the write.
Let's add additional reading after writing to solve this problem.

Jira: VLK-45500
HSDES: 18028096950

Signed-off-by: Piotr Piórkowski <piotr.piorkowski@intel.com>
Acked-by: Tomasz Lis <tomasz.lis@intel.com>