drm/i915/gt: Wrap i915_ggtt_pin() to always apply the GGTT pin bias
The low addresses are not always accessible during execution on the
engine, for example, the GuC aliases both the low and high address
ranges for its own use. For this we have i915_ggtt_pin_bias() to avoid
the aliased ranges, however some callers neglected this. In particular,
the timeline could end up at offset 0, preventing the breadcrumb writes
and the context appearing hung.

Fixes: 52954edd1f70 ("drm/i915: Allocate a status page for each timeline")
Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Cc: <stable@vger.kernel.org> # v5.1+