drm/i915/guc: Relieve interrupt pressure on the CT queues before sending
In order to send a message, we need to make sure there is enough space
for the ack/reply from the guc in our recv CT buffer. However, under
sufficient interrupt pressure, draining the CT recv queue can be delayed
and may not happen before we saturate the send queue leading to us
declaring a CT communication stall and wedging the driver.

To mitigate the impact of a delay in interrupt processing, check for an
reply before performing the next send.

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>