INTEL_DII: drm/i915/svm: Migrate one page on one page fault
Previously we migrate a whole vma on a gpu page fault. For vma that
covers big address range but GPU only access it sparsely, this is
not efficient. Change the scheme to one page per page fault for now.

However this could be not efficient if GPU actually access the whole
range of the vma. So it might be helpful to introduce a parameter
for user to select the migration granularity. We will decide this
when we have more performance data.

One byproduct of this change is, it fixes an migration out of bound
issue. When vma's size is bigger than 2M, Linux kernel append an
extra page at the end of vma for meta data purpose. This page doesn't
belong to user space and shouldn't be migrated/mapped to GPU. The
migration end address is carefully calculated to avoid this issue.

Signed-off-by: Oak Zeng <oak.zeng@intel.com>
Reviewed-by: Brian Welty <brian.welty@intel.com>