INTEL_DII/UAPI: drm/i915/mmap: Add mmap support for PCI memory barrier
In order to avoid having userspace to use MI_MEM_FENCE,
we are adding a mechanism for userspace to generate a
PCI memory barrier with low overhead (avoiding IOCTL call).

This is implemented by memory-mapping a page as uncached
that is backed by MMIO on the dGPU and thus allowing userspace
to do memory write to the page without invoking an IOCTL which
would cause a context switch. We are selecting the MMIO so
that it is not accessible from the PCI bus so that the MMIO
writes themselves are ignored, but the PCI memory barrier will
still take action as the MMIO filtering will happen after the
memory barrier effect.

When we detect special defined offset in mmap(), We are mapping
4K page which contains the last of page of doorbell MMIO range
to userspace for same purpose.

V3:
  - Handle PROT_EXEC - vlk-46713
V2:
  - Move #define to i915_drm_prelim.h - adixit
  - Add UAPI and needs IGT - Matt

Bug-id: VLK-39126
Bspec: 53080
Cc: Chris Wilson <chris.p.wilson@linux.intel.com>
Cc: Michael J. Ruhl <michael.j.ruhl@intel.com>
Cc: Nirmoy Das <nirmoy.das@intel.com>
Cc: Matt Roper <matthew.d.roper@intel.com>
Cc: Ashutosh Dixit <ashutosh.dixit@intel.com>
Signed-off-by: Tejas Upadhyay <tejas.upadhyay@intel.com>