PF: Force GuC VGT mode after reset
Initially we were forcing GuC GVT mode just for the selftests,
but due to some GuC limitations we extended that to whole life
of the PF. But that mode change was not aplied after GT/GuC reset.
Fix that and do it in the single patch that will stay with us for
longer time as we don't expect GuC to drop the idea of separate VGT
mode of operation.

References: VLK-42956
Signed-off-by: Michal Wajdeczko <michal.wajdeczko@intel.com>
Acked-by: Piotr Piórkowski <piotr.piorkowski@intel.com>
(cherry picked from commit 984ed0531014091ca5d43e7928b239d2242473e7)