Fixed CCS mode vs clear-on-free
For fixed CCS mode on PVC, we do not allow changing the CCS
configuration while one user engine is still active, including any of
the BCS engines (this is keep the configuration valid while the client
is still processing and preparing future operations). Clear-on-free
changes the timing of the activity on the context. Whereas before we did
not wake the engines until demand paging in the new buffers, the engine
remain active from the previous client as we compare the CCS
configuration required for the next client. It's a very subtle change,
unlikely to affect anything in practice!

To accommodate one negative IGT (i915_query/engines), consider the
validity of the execbuf before changing the CCS mode. And when we do try
to change the CCS mode, remember the current register value to avoid
switching unnecessarily.

The important changes are:

- moving the really expensive assert into the debug only code (~4ms
  executon stall), and only doing it on a change (before we checked and
  only warned in debug)

- only changing the CCS_MODE when the register value must change.

- doing the checks after confirming the execbuf is valid for execution,
  so skipping the issue where we check early and see that clear-on-free
  is still active from the previous test run.

Jira: vlk-47107
Reviewed-by: Stuart Summers <stuart.summers@intel.com>
(cherry picked from commit 10579f5b1743e8dd11229cec4b26a0a45b8ad49f)