Local memory test
Currently we do a write of 0x0 to every bytes of free lmem in order to
trigger patrol scrub and MCE if the HW detects any memory errors.  After
discussion with SV, it seems the current set of issues is only
reproducible by patrol scrub which does not actually run on-demand, and
instead is run at an interval of once per 24 hours. The MCE is
apparently hit consistently during each patrol scrub (i.e. once per 24
hours), and there is a desire to increase the frequency of occurrence
such that each card in the Argonne cluster can be triggered if it is
exposed to an issue with the HBMs in a short amount of time (matter of
seconds rather than 24 hours).

The patrol scrub interval is programmable via registers only
accessible via ITP. By programming these to an interval as short as 7 or
8 seconds, SV is able to consistently reproduce the HBM errors.  However
given ITP is not available at scale, it is desirable to have a way to
duplicate the patrol scrub sequence from the host.

Add a new lmemtest which issues a read/modify/write using multiple
different bit patterns across all of local memory. We want to allow this
to be exercised at runtime and so need not only test the free local
memory, but even the memory in use. For that, we interrupt the
execution, copy the active memory to a temporary buffer, perform, the
destructive write testing, then restore the user contents.

Given we don't know exactly why only the patrol scrub currently triggers
the HBM MCEs, it's possible this test will need to be tweaked in the
future. However, given the current level of SV testing, this is the most
likely candidate and allows us to scrub the HBMs easily, ensuring all of
local memory is healthy and a user should not hit this class of issue at
runtime.

As a difference to the previous version that just cleared all memory,
this requires a write to trigger a new scan. This is so that we do not
trigger a full scan accidentally on the simulator.

$ time echo 0 | sudo tee /sys/kernel/debug/dri/0/lmemtest
0

real	0m2.246s # pvc
user	0m0.011s
sys	0m0.001s

$ sudo cat /sys/kernel/debug/dri/0/lmemtest
0x0000000000000000

It takes around 2s on pvc to scan 128GiB of memory, performing multiple
passes in each scan.

v2: Manual runalone to prevent concurrent execution
v3: Rebase onto clear-on-free, use Stuart's more detailed commitmsg
v4: Use ppHWSP for atomics, as that is in lmem (required for pvc+)
v5: Fix randomly sampling out-of-bounds

Jira: vlk-46150, vlk-45634
Cc: Stuart Summers <stuart.summers@intel.com>
Reviewed-by: Stuart Summers <stuart.summers@intel.com>
(cherry picked from commit 3d38bd50b2961099f2da0b19be4bbba0f380fa4d)