Deferred clear awaits
Deferred clears set the migrate fence which must be waited upon before
acccessing the object. Normally awaits are peformed via the vma (which
is chained to the migration fence) but for completeness, await on the
migration fence directly, in case the kernel tries to bypass fences.

Catch a couple of klockworks in the process.

Jira: vlk-47064
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>
(cherry picked from commit 6ea483197cd88318d50bd61e5e39778aca91933f)