Backpressure for clear-on-free
Don't block waiting for outstanding clears, these may be blocked behind
userspace, for opportunistic clear-on-free. In particular, it is not
expected for the the free to block, which in turn may cause another
syscall to block (through i915_gem_flush_freed_objects), and so may add
an unexpected blocking ioctl to userspace. If we back off and skip the
clear-on-free when the clear context is overloaded with work, those
clears are deferred to idle, or to clear-on-alloc under memory pressure.

Jira: vlk-47109
Reviewed-by: Stuart Summers <stuart.summers@intel.com>
(cherry picked from commit 9edcefc9f05290204323e17cb1c3c80e6a33ecdd)