Rebalance parallel eviction
A dilemma in parallel eviction scenarios is that we need to retry even
if we find enough objects this pass as we may lose the race against a
concurrent eviction. However, if we are the only evictor, we do not want
to keep repeating ourselves.

Jira: vlk-47344
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>
(cherry picked from commit c100777754e364542adf0fa39888cb8f6f00ca5f)