Fix iov_ggtt@igt_vf_other_ggtt for TGL/ADL
Originally, the Bspec [1] stated that PTE readings outside the range of
a given VF would result in a reading of 0x0.
In practice, it turned out that on TGL/ADL this value is
0xffffffffffffffff.
Discussion at HSDES [2] revealed that this is a bug in the specification
and Bspec has been updated.
Let's update the test to match what is currently in Bspec.

References: HSDES:18027049632 [1], Bspec:52395 [2]

Signed-off-by: Piotr Piórkowski <piotr.piorkowski@intel.com>
Acked-by: Tomasz Lis <tomasz.lis@intel.com>
Acked-by: Michal Wajdeczko <michal.wajdeczko@intel.com>
(cherry picked from commit f0d25c87eb35c94b36d70b2504f3e044bd04d1bf)