Always wait for clears before installing sriov LMTT
Since deferred clears, it is possible for the clear-on-use to be
asynchronous and so it is mandatory to wait for the clear to be
completed before accessing the backing store directly. With the advent
of clear-on-free, we then try to clear all pages upon release
potentially forcing the next alloc to be asynchronous even when the
caller hints to use the cpu, and thus /must/ wait. As such is the case
for LMTT (sriov).

Include all the hidden chickens for easier deferred clear feature
testing.

Jira: vlk-47452
(cherry picked from commit 92c5c9fc72a65b81518767540b965c1f792ebfd7)