Avoid clear_blt for synchronising too-small chunks
In vlk-47452, we observed that using XY_SURF_CTL_COPY on chunks smaller
than 64K would result in alignment errors and CAT faults (dg2 with
flat-ccs). By restricting clear_blt to only operate on
I915_BO_ALLOC_USER, we thereby prevent trying to clear smaller 4K chunks
(used for pagetables).

However, we are still using clear_blt() to create a synchronisation
fence for pending dirty blocks, even for !I915_BO_ALLOC_USER which means
we may still attempt to use the blitter with those same chunks. Wrap
those fences into a separate dma-fence outside of clear_blt to eliminate
the final source of bad chunks.

The intent to always use clear_blt() to generate a i915_request was to
ensure we had full PI. For too small chunks, we should never be clearing
them in the first place and wish to immediately use them (pagetables) on
the cpu so are not subject to priority-inheritance concerns along the
pipeline.

Jira: vlk-47452
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>