Use bcs0 for clear-on-idle
Promote clear-on-idle to use the much faster BCS0 (rather than BCS8 on
PVC). Since clear-on-idle should be running when there is no contention
with userspace, stealing the high bandwidth engine for our background
clears should not be an issue, and thus we want to "race to idle" and
clear as much memory as possible before userspace wakes up.

Reviewed-by: Brian Welty <brian.welty@intel.com>
Reviewed-by: Andrzej Hajda <andrzej.hajda@intel.com>
(cherry picked from commit b5e2852ec3a9975ab991be44913764fdb69364ff)