Preparatory patches for uptr/smem pipelining
In the next series, we will begin utilising pipelined acquisition of
system memory resources (userptr and shmemfs). To easily accommodate
the asynchronous acquisition of that backing store while fitting into
the existing driver, we will be using a placeholder scattergather list
and populate it later with the result of the asynchronous gather. This
means at the time of i915_gem_object_set_pages() the sg table is
incomplete and we do not have any knowledge of the physical layout of
the pages. So this series removes our use of any prior knowledge of the
sglist, deriving the physical properties of the those elements as we need
them, and enables inplace manipulation of the sg_table.

Jira: vlk-42294, vlk-47071 # prep
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>
(cherry picked from commit 174d98a9b51c621e08819eeae7eaa309d639eb7c)