Extend NULL PTE support to gen12+ (ats-m)
On dg2, we need a 64K scratch page for each context/vm. This imposes
extra mempressure on dg2/ats-m as we frequently cannot find a contiguous
64K page after the system has been running for some time. However, since
gen12 supports a NULL PTE we can use that to drop scratch writes without
requring the page to be allocated. (On earlier gen, we could mark the
page as read-only, and so reuse the same scratch pages for all.)

Reviewed-by: Bruce Chang <yu.bruce.chang@intel.com>
(cherry picked from commit fb3c46e9f473a676523bb224607b5a53a37a1a65)