GuC 70.7.0 + WA 14018913170 for DG2/PVC
GuC 70.7.0 introduces support for WA 14018913170, with a new flag added
to the interface to be used by the driver to enable the WA on the
platforms it applies to. Unfortunately, even though the HW team has
confirmed that the WA applies to most modern platforms, the HSD status
is still on "pending review" for the majority on them. Given that we
have an open issue that is fixed by this WA (VLK-45273), the PM
requested the WA to be enabled anyway for DG2 and PVC, even if the HSD
is not confirmed. Other platforms will be added once the HSDs are sorted
out.

GuC 70.7.0 also uses a new HuC auth keys for MTL/LNL/BMG, so we can't
switch those until the HuC blobs using the new keys are ready.

Changelog since 70.6.8:

Feature/Refactor/Tuning:
 o Refactor: Add New HuC Authentication Keys
 o Feature: Add WA for TSC not synched at RC6 exit
BugFix/Workaround:
 o Bugfix: Protect PF2VF Relay with VF Exception TryCatch
 o Bugfix: Cache Flush Before HuC Signature Buffer Read
 o Bugfix: Invalidate GuC TLB before Updating Attribute Table Buffer
 o BugFix: Allow KMD Initiated P2I on Faulted Engine
 o BugFix: NPK logging on emulation no longer functions

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Reviewed-by: Alan Previn <alan.previn.teres.alexis@intel.com>