Don't keep a persistent error on the vma for interrupted construction
If we encounter an error during construction of the vma-bind work, such
as hitting a signal interrupt, we immediately abort the vma-bind and
report the error back to the user. However, we would still mark the
interrupted work as resulting in a persistent VMA error. Simply, ignore
the error during the setup to allow the VMA to be rebound again by the
user.

Jira: vlk-48657
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>