Remove GT active deferral of vma close
In "Decouple GT parking and vma close", there was a race condition in
vma lookup by execbuf and closing that was temporarily closed by
deferring the closure until the GT was idle (as was previously the
case). However, this deferral make last for an eternity effectively
leaking memory for an always active system, such as a high tenancy
system. Let's see if the root cause was truly found and fixed by
switching back to a pure time based vma open cache.

The race being papered over was found by CI, so hopefully the same
timing and coverage still exists and we can detect if this approach will
fail.

Jira: gsd-5291 / vlk-48924
Fixes: "Decouple GT parking and vma close" [DII_6777]
Acked-by: Fei Yang <fei.yang@intel.com>