INTEL_DII: drm/i915/iaf: Address two IAF Jiras
For IAF, Netlink is enabled on module load and can be accessed before
module unload has completed.  This leaves a window for users to access
freed memory.  Address the race.

Jira: VLK-47900

i915 GEM objects are cleaned up in a lazy fashion.  This can impact
device removal if a mapping ref (IAF) has been incremented.

Using the child to device message API, add a message to notify the
parent that the child is going to be removed, in case the parent
can do any pre-cleanup.

Jira: GSD-3189

Reviewed-by: David Kershner <david.kershner@intel.com>
Acked-by: Gary S Leshner <gary.s.leshner@intel.com>
Signed-off-by: Michael J. Ruhl <michael.j.ruhl@intel.com>