Remove GT active deferral of vma close
In "Decouple GT parking and vma close", there was a race condition in
vma lookup by execbuf and closing that was temporarily closed by
deferring the closure until the GT was idle (as was previously the
case). However, this deferral make last for an eternity effectively
leaking memory for an always active system, such as a high tenancy
system. Let's see if the root cause was truly found and fixed by
switching back to a pure time based vma open cache.

v2: Close the race between lookup+open and close with a semaphore. This
focuses on the execbuffer path, i.e. where vma are lookup+open, as the
race should only be present here.

Jira: gsd-5291 / vlk-48924
Fixes: "Decouple GT parking and vma close" [DII_6777]
Acked-by: Nirmoy Das <nirmoy.das@intel.com>